rename machinedeploymentazs resource to tcnpazs
