+++ service/controller/clusterapi/v29/resource/tcnpazs/create.go
@@ -1,4 +1,4 @@
-package machinedeploymentazs
+package tcnpazs
 
 import (
 	"context"
