+++ service/controller/clusterapi/v29/resource/tcnpazs/delete.go
@@ -1,4 +1,4 @@
-package machinedeploymentazs
+package tcnpazs
 
 import (
 	"context"
