+++ service/controller/clusterapi/v29/resource/tcnpazs/resource.go
@@ -1,13 +1,12 @@
-// Package machinedeploymentazs implements a resource to gather all private
-// subnets for the configured availability zones of a node pool. Like the
-// clusterazs resource, we need logic to take the node pool subnet allocated by
-// the ipam resource and split it according to the configured availability
-// zones. We then have 1, 2 or 4 private subnet CIDRs we put into the controller
-// context for further use in the tcnp resource. Note that the availability
-// zones of a node pool cannot be updated upon creation due to the network
-// splitting. In order to change availability zones one must delete and create
-// node pools accordingly.
-package machinedeploymentazs
+// Package tcnpazs implements a resource to gather all private subnets for the
+// configured availability zones of a node pool. Like the clusterazs resource,
+// we need logic to take the node pool subnet allocated by the ipam resource and
+// split it according to the configured availability zones. We then have 1, 2 or
+// 4 private subnet CIDRs we put into the controller context for further use in
+// the tcnp resource. Note that the availability zones of a node pool cannot be
+// updated upon creation due to the network splitting. In order to change
+// availability zones one must delete and create node pools accordingly.
+package tcnpazs
 
 import (
 	"github.com/giantswarm/microerror"
@@ -16,7 +15,7 @@ import (
 )
 
 const (
-	Name = "machinedeploymentazsv29"
+	Name = "tcnpazsv29"
 )
 
 type Config struct {
